requirements.txt
pyyaml
pprint
datetime
caldav
icalendar
argparse
urllib